machine-learning/TourRecommendation/requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-aiplatform
protobuf